backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.54.0
boto3==1.29.0
pandas==2.1.3
pyarrow==14.0.1
python-multipart==0.0.6
pydantic==2.5.0

